requirements/base.txt
Django==2.1.1
Pillow==5.2.0
dj-database-url==0.5.0
django-environ==0.4.5
django-markdownx==2.0.24
django-simple-captcha==0.5.9
requests==2.19.1
whitenoise==4.1
